[spmd] use ops db for testing
This PR use ops db to test all the ops that distributed tensor supports.
This is crucial for us as we are trying to increase the coverage of
different operators, and we can't write every tests for so many opereators.

So we pull in the operator tests from the core repo and use that to test
all the opereators to reduce the burden.