Implement shape ops (#324)
Implements reshape / view / transpose / permute / expand and other ops that only change the shape and layout of the tensor without changing the element's values.

This adds support for multiple mesh dimension ; however it won't succeed if communication would be triggered by the operation. This would be the case when flattening a tensor where the tensor is shared on a flatenend dimension other than the first. We will need to tackle that subsequently.