[dtensor] support implicit broadcasting for pointwise ops
This PR support implicit broadcasting for pointwise ops, where it
is able to propagate the sharding of "broadcasting to a common shape"
case.

When the dimension of a input is 1, we mark it explictly during propagation,
so that the sharding can be propagated without trying to reshard
i.e.
input0 (8, 4, 8), input 1(8, 1, 8), if we shard on input0 dim 1, it could
directly propagate the sharding without leeting input 1 be also shard on dim1

NOTE that since the op rules might be used by some chained suggestions (i.e.
addmm or baddbmm chained suggestion), one spec might be the output of a sharding
propagation, so it must have shape there to validate if there's singleton dimension.
As a result, this PR forces sharding propagation rules to generate correct output shape