[dtensor] add reshard suggestion to dfferent sharding on same mesh dim
This add reshard suggestion for the case where different inputs shrading
on the same mesh dim. In order to perform local computation, we need to
all_gather one of the inputs on that mesh dimension. We uses a simple
heuristic to determine which input to all_gather (i.e. min comm volume)

ghstack-source-id: bc10f8aa15d68f9c66e8a5c30d41a866dd45b524
Pull Request resolved: https://github.com/pytorch/tau/pull/496