Update base for Update on "[dtensor] add reshard suggestion to dfferent sharding on same mesh dim"
This add reshard suggestion for the case where different inputs shrading
on the same mesh dim. In order to perform local computation, we need to
all_gather one of the inputs on that mesh dimension. We uses a simple
heuristic to determine which input to all_gather (i.e. min comm volume)

[ghstack-poisoned]