[PT-D][Tensor Parallelism] Implement parallelize_module API and refactor _parallelize_multihead_attn (#634)
We are building a new unified high-level API so that user can directly
use it for tensor parallelism. Our tensor parallelism is built on top of
DTensor (RFC: https://github.com/pytorch/pytorch/issues/88838).

We have already built some APIs for Multihead_attention parallelism and
we now just refactor it to make sure it is working with our overall
design for tensor parallelism.