Feature run spmd - add standalone driver program for SPMD work (#661)
This PR adds a driver program, run_spmd.py, that can be used to run SPMD
directly in a configurable environment with a given sample model.
Usage is simply "python run_spmd.py" from the spmd directory. 
It sets up and runs with a configurable world size and configurable
models.
It then runs the model using both DDP and SPMD, and then does a
comparison of the grads returned via allclose as well as a two row
visual display of a grad subset.
<img width="772" alt="run_smpd_results"
src="https://user-images.githubusercontent.com/46302957/204891685-bb2d3239-f017-4376-8341-c46d59883f07.png">

By default it runs a small 2 layer linear model, with a configurable
layer size.
Also in there is the permute model subset from Rich. 

I've been using it for the fusion work and found it really helpful for
being able to quickly run and experiment/verify things. Thus wanted to
add it directly to the repo so hopefully it is useful for others by
being able to quickly spin up a given model and environment for SPMD.