Fix issue when backward graph returning sharded tensors
When SPMD is part of a larger model, backard() could output sharded
gradients (for input) as well as replicated gradients (for parameters).

This fixes a bug that was causing for non-partial/ non-replicated
outputs to disappear from the graph output.

--

Also, not necessary for this specific fix, but something that could
trigger issues down the road -- the output_specs were not reflecring the
Partial -> Replicated  resharding on the outputs. This fixes it so that
_convert_to_distributed always returns the proper output sharding for
the graph.
(One thing we could do possibly: return DTensors from SPMD if the inputs
are DTensor or something like that)