clone node offset, fix node issue from deepcopy, add autodepth finder for comm node (#671)
This PR is to update two aspects:
1 - rollback the changes that assumed clone nodes would not be present 
2 - fix a tricky bug from using deepcopy for the fe node list. 
3 - adds automatic offset finder for the comm node to replace the
hardcoded offset within the FusionElement node list.

Note - unit test is pending, but needs some further integration work for
the new api/distributed breakout. Wanted to get the above fixes in
directly.
This file with changes is not live/hooked in yet to be clear.