Feature overlap spread - moves all comm calls close to source for higher comm/compute overlap (#673)
This PR implements overlap spreading of the given all_reduce calls in
the backward graph.
Currently all comm calls end up in a big pile at the end of any given
graph. As a result, we have a large block of communication at the end of
each iteration, with 0.0 overlap with compute (vs an ideal of 90%+).

This PR checks for the source node of each comm_call, and then moves the
upper half of the SPMD comm calls to be next to this source in the
graph. This ensure that each all_reduce is triggered as soon as the
gradient tensor is ready.
The wait_comm and get items remain at the bottom of the graph to allow
for maximum time of receipt.

Unit test:
This PR adds a first unit test for the overlap by running with and
without overlap (controlled by the SPMD param, optimize_first_iter on or
off) and compares the grads to ensure identical results with the overlap
spreading.

Second unit test will come as next PR - this one will do a graph
inspection of the comm and wait node locations, but will do that as
additional PR to get this feature in.

What a graph looks like after the overlap pass:
<img width="1154" alt="working_transform"
src="https://user-images.githubusercontent.com/46302957/205456268-7fba570b-bf77-4a07-858c-d54b0d144b89.png">