update to latest graph_utils, fix comm index, update run_spmd to latest SPMD params (#686)
This PR has three small updates:
1 - The comm index for FE units was not being set in graph_info after
the graph_utils update. This meant the entire comm block was being moved
in comm_overlap.
This updates graph_utils to return this index and do it from a top down
perspective so the index is correct, vs the reversed order they are now
searched in, and store it in the global GraphInfo. (also adds an _ for
third unpack in distribute to match the signature).

2 - Updates run_spmd to use the new graph_optimizations param vs relying
on the previous first_run_iter to trigger the optimization.

3 - adds an initialized index and an assert to print debug info if we
fail to located comm_section during comm_overlap. (fixes issue Anjali
hit when comm search failed due to missing tensor attr).