Fix distributed.py when op output is a list
When the output  of an op is a list we need to hack up to "recompose"
multiple tensor outputs back to the original list. (because later this
list will be getitem'd into by the subsequent operators).

I believe later on we should redo this in a better way, but for now this
should fix our layer_norm_backward that returns a tuple of 3 tensors.