Ring fusion - adds overlapped comm fusion for SPMD using a ring buffer (#697)
This PR adds overlapped fusion using a ring buffer.
It has two params:
a - fusion_policy which is the total number of comm calls to fuse (i.e.
60 = fuse 60 calls into 1)
b - buffer_policy, which controls how many unique buffers to have in the
ring. Buffers are handed out in a cycle
to each respective fusion operation. 
The reason for the cycling is to avoid overwriting the buffer - we do
not mutex the buffer, and comms are asynch.
Thus we can have a situation where a new fusion starts writing into an
existing buffer before the outputs from the previous call are copied
out.