Remove requirement for loss spec (#709)
Previously, we require users to provide a one-hot map to specify which
output value is the loss.
For example, for T5:
```
    output_loss_value_spec = {
        "loss": True,
        "logits": False,
        "encoder_last_hidden_state": False,
    }
```
This requires users to have a good understanding of the model.

With this PR, we try to free users from such specification. We do so by
automatically searching for "loss" in the output dict. (This is the most
prevalent format for Hugging Face models.)

Previously, `output_loss_value_spec=None` has been used to indicate the
inference mode. Now, to do the same, users can set the model to evaluate
mode before passing it to PiPPy. That is,
```
model.eval()
pipe = Pipe.from_tracing(model, ...)
```