Fix issue that PiPPy does not generate backward pass when there are multiple output values (#738)
Issue was reported by user when trying the ResNet example:
[Pippy I can’t see backward
pass](https://discuss.pytorch.org/t/pippy-i-cant-see-backward-pass/170630)

### Explanation of the Issue:
When the output is in tuple form and there other values than loss in the
tuple, e.g. `return output, loss`, PiPPy could not figure out which
field is the loss, and hence did not generate the backward pass.

### Fix to the ResNet example:
The return needs to be in a form of dict, e.g. `return {"output":
output, "loss": loss}`, so that PiPPy can automatically locate the loss
field by looking at the keyword.

### Fix to PiPPy's IR:
We cannot assume that there is only one output value (loss) when the
model is of the general `LossWrapper` class. Such assumption is only
valid for the `TrivialLossWrapper` class.