Setting the default value of PIPPY_PIN_DEVICE to 0 (#766)
Previously this env was added to work around an issue that each RPC
process creates an extra CUDA context on device 0. This issue may have
been caused by RPC not automatically pinning spawned worker threads to
same CUDA device as the main thread. So pinning each RPC process to one
device would avoid the issue.

The above issue seems to be gone now. Hence, we are setting the default
value of `PIPPY_PIN_DEVICE` to 0 now.