combine index-file-saving and params-saving into one API (#823)
## Description

bring together `_save_index` and `_save_params` into one function
`save_checkpoint`. The entire module is used to write the index file,
only in rank 0, while submodule parameters are saved to a file in each
rank. The associated ckpt test is updated to reflect this change.

## Type of change

Please delete options that are not relevant.

- [x] New feature (non-breaking change which adds functionality)
- [x] This change requires a documentation update

## Feature/Issue validation/testing

- [x] Test LocalIndexMetadataTest

## Checklist:

- [x] Have you added tests that prove your fix is effective or that this
feature works?
- [x] Has code been commented, particularly in hard-to-understand areas?
- [x] Have you made corresponding changes to the documentation?

---------

Co-authored-by: Eddy <ogolaedd@meta.com>