Update setup.py to include the missing dependency (#953)
This missing dependency results in an incomplete list of deps being
included in the METADATA part of the wheel, leading to pip-compile like
tools failures to resolve all necessary dependencies if not built from
source.

## Description

Please read our
[CONTRIBUTING.md](https://github.com/pytorch/PiPPy/blob/main/CONTRIBUTING.md)
prior to creating your first pull request.

Please include a summary of the feature or issue being fixed. Please
also include relevant motivation and context. List any dependencies that
are required for this change.

Fixes #(issue)

## Type of change

Please delete options that are not relevant.

- [ ] Bug fix (non-breaking change which fixes an issue)

## Feature/Issue validation/testing

N/A


## Checklist:

N/A