Remove incorrect assertion on loss shapes
output and target shapes are expected to differ.

It might be possible to craft some shape-checking logic here, but it
would need to account for various loss functions and be general, not
sure its feasible. Removing for now to unblock.

(using cross-entropy loss in this case- see docs about expected shapes
https://pytorch.org/docs/stable/generated/torch.nn.functional.cross_entropy.html)

ghstack-source-id: 265b02d33ba796a5b6e99c81fe291c295ccae61f
Pull Request resolved: https://github.com/pytorch/pippy/pull/1025