Update on "Use context manager to replace spec arguments of pipeline API"
Removing these arguments from the `pipeline` API (which is used to create a `Pipe`):
```
def pipeline(
     module: torch.nn.Module,
     num_chunks: int,
     example_args: Tuple[Any, ...],
     example_kwargs: Optional[Dict[str, Any]] = None,
     split_policy: Optional[Callable[[fx.GraphModule], fx.GraphModule]] = None,
-    args_chunk_spec: Optional[Tuple[Any, ...]] = None,
-    kwargs_chunk_spec: Optional[Dict[str, Any]] = None,
-    output_chunk_spec=None,
-    constraints: Optional[List[Constraint]] = None,
) -> Pipe:
```

`args_chunk_spec` and `kwargs_chunk_spec` are now supported via a context manager:
```
# Numbers here are tensor chunking dimensions
with ArgsChunkSpec((0, 0, 1)), KwargsChunkSpec({"id": 0}):
    pipe = pipeline(model, num_chunks, example_args, example_kwargs)
```

For most cases, users don't need to provide the chunking dimensions, the default dimension of 0 (batch dimension) will be used. So the API would be as simple as:
```
pipe = pipeline(model, num_chunks, example_args, example_kwargs)
```

`output_chunk_spec` is gone because I plan to move it to PipelineSchedule -- user tells us there how they want the output to be presented. Also because IR does not need output_chunk_spec.

`constraint` is gone because it is gone from torch.export's API. (Replaced by something else). And we'd never truly have support for constraint in pippy. We can add support for the something else in future.




[ghstack-poisoned]